+++ code/simple_cube.py
@@ -81,7 +81,7 @@ class Quaternion:
                          + prod[2, 0] + prod[3, 1]),
                         (prod[0, 3] + prod[1, 2]
                          - prod[2, 1] + prod[3, 0])],
-                       dtype=np.float,
+                       dtype=np.float32,
                        order='F').T
         return self.__class__(ret.reshape(return_shape))
